[Fix] docker (#16)
* Update build paths for API server binary

Changed the build commands in Dockerfile and Makefile to use the new path './cmd/api' in Dockerfile and './api/cmd' in Makefile for building the apiserver binary. This aligns the build process with the updated project structure.

* Remove extra blank line in CI workflow

Deleted an unnecessary blank line before the Gosec Security Scanner step in the GitHub Actions CI workflow for improved formatting.